refactor: edit/addView 상속 통해 구현
ProductEditableView를 이용해서 공통 부분을 구현했음
